added automo concept, aligned the first APIs with the mongo driver 3.x
